# api/openapi.yaml
openapi: 3.0.1
info:
  title: MERN   API
  version: 1.0.0
  description: API endpoints [social media app, classroom app]
  contact:
    email: fmg3ckali@gmail.com
    name: Nathaniel Cobbinah
    url: https://shy-lime-dibbler-garb.cyclic.app/
  license:
    name: MIT
    url: https://choosealicense.com/licenses/mit/
paths:
...
  /api/courses/by/{userId}:
    post:
      summary: 'verify user is an educator before creating course'
      description: >-
        verifies that the current user is an educator, and then creates a new course with the course data 
      operationId: createCourse
      parameters:
        - name: userId
          in: path
          description: The userId of the user.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: A new course object
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Classroom'
        required: true
      responses:
        '200':
          content:
            application/json:
              examples:
                User who is an educator created course saved Successfully:
                  value: |-
                    { 
                     "name":"John Smith",
                     "description":"john@smith.info",
                     "image": "wwwwwwwwwwwwwwwww"
                     "category": "xxxxxsewsdsds",
                     "published": "sssssssse",
                     "instructor": "ssssssssss",
                     "updated":"ssssssssssss",
                     "created":"sssssssssssssss",
                     "lessons":[
                        {
                          "title":"wwwwwwwwww",
                          "content":"eeeeeeee",
                          "resource_url":"eeeeeeeeeeeeee"
                        }
                     ]
                    }
          description: Creates a new post record by user
        default:
          description: Default error sample response
      tags:
        - Course
      security:
        - bearerAuth: []
    
    get:
      summary: 'returns the list of courses by a given instructor'
      description: ''
      operationId: listByInstructor
      parameters:
        - name: userId
          in: path
          description: The userId of the user.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              examples:
                list of courses by a given instructor retrieved successfully:
                  value: |-
                    [
                       { 
                       "name":"John Smith",
                       "description":"john@smith.info",
                       "image": "wwwwwwwwwwwwwwwww"
                       "category": "xxxxxsewsdsds",
                       "published": "sssssssse",
                       "instructor": "ssssssssss",
                       "updated":"ssssssssssss",
                       "created":"sssssssssssssss",
                       "lessons":[
                          {
                            "title":"wwwwwwwwww",
                            "content":"eeeeeeee",
                            "resource_url":"eeeeeeeeeeeeee"
                          }
                       ]
                      }
                    ]
          description: Retrieved list of courses by user who is an educator
        default:
          description: Default error sample response
      tags:
        - Course
      security:
        - bearerAuth: []
  /api/courses/photo/{courseId}:
    get:
      summary: Returns the photo associated with a specific course
      description: ''
      operationId: courseImage
      parameters:
        - name: courseId
          in: path
          description: Image associated with a course.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: Image file for course
          content:
            image/*:
              schema:
                type: string
                format: binary
        default:
          description: Default error sample response
      tags:
        - Course
  /api/courses/{courseId}:
    get:
      summary: Returns individual course specific details
      description: ''
      operationId: readCourseInfo
      parameters:
        - name: courseId
          in: path
          description: Returns individual course specific details.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          description: Default sample response
          content:
            application/json:
              examples:
                User who is an educator created course saved Successfully:
                  value: |-
                    { 
                     "name":"John Smith",
                     "description":"john@smith.info",
                     "image": "wwwwwwwwwwwwwwwww"
                     "category": "xxxxxsewsdsds",
                     "published": "sssssssse",
                     "instructor": "ssssssssss",
                     "updated":"ssssssssssss",
                     "created":"sssssssssssssss",
                     "lessons":[
                        {
                          "title":"wwwwwwwwww",
                          "content":"eeeeeeee",
                          "resource_url":"eeeeeeeeeeeeee"
                        }
                     ]
                    }
        default:
          description: Default error sample response
      tags:
        - Course   
        
    put:
      summary: ' allows an existing course to be updated if the user who is making the request is the authorized instructor of the given course'
      description: ''
      operationId: updateCourseInfo
      parameters:
        - name: courseId
          in: path
          description: The courseId for a given course.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: required data for updating a given course
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CourseWithLessonsUpdate'
        required: true
      
      responses:
        '200':
          description: course content updated successfully
        default:
          description: Default error sample response
      tags:
        - Course
      security:
        - bearerAuth: []
        
    delete:
      summary: takes a request to delete a specified course from the database
      description: >-
        This DELETE route takes the course ID as a URL parameter and checks if the current user is signed in and authorized to perform this delete, before proceeding to delete the course from the database
      operationId: removeCourse
      parameters:
        - name: courseId
          in: path
          description: The courseId of a course.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              examples:
                Course Record deleted successfully:
                  value: |-
                    [
                       { 
                       "name":"John Smith",
                       "description":"john@smith.info",
                       "image": "wwwwwwwwwwwwwwwww"
                       "category": "xxxxxsewsdsds",
                       "published": "sssssssse",
                       "instructor": "ssssssssss",
                       "updated":"ssssssssssss",
                       "created":"sssssssssssssss",
                       "lessons":[
                          {
                            "title":"wwwwwwwwww",
                            "content":"eeeeeeee",
                            "resource_url":"eeeeeeeeeeeeee"
                          }
                       ]
                      }
                    ]
          description: Removes a course from the database
        default:
          description: Default error sample response
      tags:
        - Course
      security:
        - bearerAuth: []
  /api/courses/published:
    get:
      summary: retrieve the list of published courses from the database
      description: ''
      operationId: listPublished
      responses:
        '200':
          description: Default sample response
          content:
            application/json:
              examples:
                retrieved the list of published courses from the database successfully:
                  value: |-
                    [
                      { 
                       "name":"John Smith",
                       "description":"john@smith.info",
                       "image": "wwwwwwwwwwwwwwwww"
                       "category": "xxxxxsewsdsds",
                       "published": "sssssssse",
                       "instructor": "ssssssssss",
                       "updated":"ssssssssssss",
                       "created":"sssssssssssssss",
                       "lessons":[
                          {
                            "title":"wwwwwwwwww",
                            "content":"eeeeeeee",
                            "resource_url":"eeeeeeeeeeeeee"
                          }
                       ]
                      }
                    ]
        default:
          description: Default error sample response
      tags:
        - Course 
  /api/courses/{courseId}/lesson/new:
    put:
      summary: ' while it is still unpublished, the instructor will be able to add a lesson to a course'
      description: ''
      operationId: newLesson
      parameters:
        - name: courseId
          in: path
          description: The courseId for a given course.
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: required lesson info data for updating a given course
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lesson'
        required: true
      responses:
        '200':
          description: course content updated successfully
        default:
          description: Default error sample response
      tags:
        - Course
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Classroom:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        image:
          type: string
          format: binary
    CourseWithLessonsUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        published:
          type: boolean
        image:
          type: string
          format: binary
        lessons:
          $ref: '#/components/schemas/Lesson'
    Lesson:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        resource_url:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT